Programme of International Student Assessment (PISA) results from the United States in 2009.

Student results from the 2009 Programme of International Student Assessment (PISA) as provided by the Organization for Economic Co-operation and Development (OECD). See https://www.oecd.org/pisa/ for more information including the code book.

Format

a data frame with 5,233 rows and 65 columns.

CNT

Country

SCHOOLID

SchoolID

ST01Q01

Grade

ST04Q01

Sex

ST05Q01

Attend

ST06Q01

Age

ST07Q01

Repeat

ST08Q01

At home mother

ST08Q02

At home father

ST08Q03

At home brothers

ST08Q04

At home sisters

ST08Q05

At home grandparents

ST08Q06

At home others

ST10Q01

Mother highest schooling

ST12Q01

Mother current job status

ST14Q01

Father highest schooling

ST16Q01

Father current job status

ST19Q01

Language at home

ST20Q01

Desk

ST20Q02

Own room

ST20Q03

Study place

ST20Q04

Computer

ST20Q05

Software

ST20Q06

Internet

ST20Q07

Literature

ST20Q08

Poetry

ST20Q09

Art

ST20Q10

Textbooks

ST20Q12

Dictionary

ST20Q13

Dishwasher

ST20Q14

DVD

ST21Q01

How many cellphones

ST21Q02

How many TVs

ST21Q03

How many computers

ST21Q04

How many cars

ST21Q05

How many rooms bath or shower

ST22Q01

How many books

ST23Q01

Reading enjoyment time

ST31Q01

Enrich in test language

ST31Q02

Enrich in mathematics

ST31Q03

Enrich in science

ST31Q05

Remedial in test language

ST31Q06

Remedial in mathematics

ST31Q07

Remedial in science

ST32Q01

Out of school lessons in test language

ST32Q02

Out of school lessons maths

ST32Q03

Out of school lessons in science

PUBPRIV

Public or private school

STRATIO

Student to teacher ratio in school

Details

Note that missing values have been imputed using the mice package. Details on the specific procedure are in the pisa.impute function in the pisa package.

References

Organisation for Economic Co-operation and Development (2009). Programme for International Student Assessment (PISA). https://www.oecd.org/pisa/